+++ web/documentserver-example/python/src/utils/users.py
@@ -91,7 +91,8 @@ descr_user_0 = [
     "Can't protect file",
     "View file without collaboration",
     "Can’t submit forms",
-    "Can't refresh outdated file"
+    "Can't refresh outdated file",
+    "Tour of tips when opening a document"
 ]
 
 USERS = [